+++ lib/ansible/plugins/connection/ssh.py
@@ -386,11 +386,11 @@ from ansible.errors import (
     AnsibleFileNotFound,
 )
 from ansible.module_utils.six import PY3, text_type, binary_type
+from ansible.module_utils.common.hashing import generate_secure_checksum
 from ansible.module_utils.common.text.converters import to_bytes, to_native, to_text
 from ansible.plugins.connection import ConnectionBase, BUFSIZE
 from ansible.plugins.shell.powershell import _parse_clixml
 from ansible.utils.display import Display
-from ansible.utils.hashing import generate_secure_checksum
 from ansible.utils.path import unfrackpath, makedirs_safe
 
 display = Display()
