+++ lib/ansible/plugins/inventory/__init__.py
@@ -14,12 +14,12 @@ from ansible.inventory.group import to_safe_group_name as original_safe
 from ansible.parsing.utils.addresses import parse_address
 from ansible.plugins import AnsiblePlugin
 from ansible.plugins.cache import CachePluginAdjudicator as CacheObject
+from ansible.module_utils.common.hashing import generate_secure_checksum
 from ansible.module_utils.common.text.converters import to_bytes, to_native
 from ansible.module_utils.parsing.convert_bool import boolean
 from ansible.module_utils.six import string_types
 from ansible.template import Templar
 from ansible.utils.display import Display
-from ansible.utils.hashing import generate_secure_checksum
 from ansible.utils.vars import combine_vars, load_extra_vars
 
 display = Display()
