+++ lib/ansible/plugins/lookup/password.py
@@ -131,12 +131,12 @@ import string
 import time
 
 from ansible.errors import AnsibleError, AnsibleAssertionError
+from ansible.module_utils.common.hashing import generate_secure_checksum
 from ansible.module_utils.common.text.converters import to_bytes, to_native, to_text
 from ansible.module_utils.six import string_types
 from ansible.parsing.splitter import parse_kv
 from ansible.plugins.lookup import LookupBase
 from ansible.utils.encrypt import BaseHash, do_encrypt, random_password, random_salt
-from ansible.utils.hashing import generate_secure_checksum
 from ansible.utils.path import makedirs_safe
