+++ test/units/executor/module_common/test_recursive_finder.py
@@ -41,6 +41,7 @@ MODULE_UTILS_BASIC_FILES = frozenset(('ansible/__init__.py',
                                       'ansible/module_utils/six/__init__.py',
                                       'ansible/module_utils/_text.py',
                                       'ansible/module_utils/common/collections.py',
+                                      'ansible/module_utils/common/hashing.py',
                                       'ansible/module_utils/common/parameters.py',
                                       'ansible/module_utils/common/warnings.py',
                                       'ansible/module_utils/parsing/convert_bool.py',
